+++ components/Tabs/tests/snapshot/TabBarView/MDCTabBarViewSnapshotTests.m
@@ -16,7 +16,7 @@
 
 #import "MDCTabBarItem.h"
 #import "MDCTabBarView.h"
-#import "MDCTabBarViewIndicatorSupporting.h"
+#import "MDCTabBarViewCustomViewable.h"
 #import "MDCTabBarViewItemView.h"
 
 /** The typical size of an image in a Tab bar. */
@@ -34,8 +34,13 @@ static const CGFloat kMinItemWidth = 90;
 /** The maximum width of a tab bar item. */
 static const CGFloat kMaxItemWidth = 360;
 
+static inline void SizeViewToIntrinsicContentSize(UIView *view) {
+  CGSize intrinsicContentSize = view.intrinsicContentSize;
+  view.bounds = CGRectMake(0, 0, intrinsicContentSize.width, intrinsicContentSize.height);
+}
+
 /** A custom view to place in an MDCTabBarView. */
-@interface MDCTabBarViewSnapshotTestsCustomView : UIView <MDCTabBarViewIndicatorSupporting>
+@interface MDCTabBarViewSnapshotTestsCustomView : UIView <MDCTabBarViewCustomViewable>
 /** A switch shown in the view. */
 @property(nonatomic, strong) UISwitch *aSwitch;
 @end
@@ -44,6 +49,10 @@ static const CGFloat kMaxItemWidth = 360;
 
 - (CGRect)contentFrame {
   return CGRectStandardize(self.aSwitch.frame);
+}
+
+- (void)setSelected:(BOOL)selected animated:(BOOL)animated {
+  // No-op since this is an example.
 }
 
 - (UISwitch *)aSwitch {
@@ -367,7 +376,7 @@ static NSString *const kItemTitleLong3Arabic = @"تحت أي قدما وإقام
 
 #pragma mark - Selection
 
-- (void)testChangingSelectedItemIgnoresSelectedImage {
+- (void)testChangingSelectedItemUsesSelectedImage {
   // Given
   self.tabBarView.bounds = CGRectMake(0, 0, 360, kExpectedHeightTitlesAndIcons);
   self.item1.image = self.typicalIcon1;
@@ -389,7 +398,7 @@ static NSString *const kItemTitleLong3Arabic = @"تحت أي قدما وإقام
   UITabBarItem *item1 = [[UITabBarItem alloc] initWithTitle:@"One" image:self.typicalIcon1 tag:0];
   UITabBarItem *item2 = [[UITabBarItem alloc] initWithTitle:@"Two" image:self.typicalIcon2 tag:1];
   self.tabBarView.items = @[ item1, item2 ];
-  [self.tabBarView sizeToFit];
+  SizeViewToIntrinsicContentSize(self.tabBarView);
   [self.tabBarView setSelectedItem:item1 animated:NO];
   [self.tabBarView setTitleColor:UIColor.blackColor forState:UIControlStateNormal];
   [self.tabBarView setImageTintColor:UIColor.blackColor forState:UIControlStateNormal];
@@ -476,7 +485,13 @@ static NSString *const kItemTitleLong3Arabic = @"تحت أي قدما وإقام
   [self generateSnapshotAndVerifyForView:self.tabBarView];
 }
 
-- (void)testChangingImageOfSelectedItemAfterAddingToBar {
+/**
+ This order of operations has no effect because UITabBarItem assigns a copy the value set for
+ @c image to @c selectedImage if it is currently @c nil. Since MDCTabBarView can only retrieve
+ images from the public API of UITabBarItem, there is a value for @c selectedImage. As a result,
+ changing the value of @c image will not update @c selectedImage.
+ */
+- (void)testChangingImageOfSelectedItemAfterAddingToBarDoesNothing {
   // Given
   self.tabBarView.bounds = CGRectMake(0, 0, 360, kExpectedHeightTitlesAndIcons);
   UITabBarItem *item1 = [[UITabBarItem alloc] initWithTitle:@"One" image:self.typicalIcon1 tag:0];
@@ -508,7 +523,7 @@ static NSString *const kItemTitleLong3Arabic = @"تحت أي قدما وإقام
   [self generateSnapshotAndVerifyForView:self.tabBarView];
 }
 
-- (void)testChangingSelectedImageOfSelectedItemAfterAddingToBarDoesNothing {
+- (void)testChangingSelectedImageOfSelectedItemAfterAddingToBarUpdatesView {
   // Given
   self.tabBarView.bounds = CGRectMake(0, 0, 360, kExpectedHeightTitlesAndIcons);
   UITabBarItem *item1 = [[UITabBarItem alloc] initWithTitle:@"One" image:self.typicalIcon1 tag:0];
@@ -691,8 +706,11 @@ static NSString *const kItemTitleLong3Arabic = @"تحت أي قدما وإقام
   [self.tabBarView setSelectedItem:item2 animated:NO];
 
   // When
-  superview.customSafeAreaInsets = UIEdgeInsetsMake(16, 44, 0, 0);
-  [self.tabBarView sizeToFit];
+  UIEdgeInsets safeAreaInsets = UIEdgeInsetsMake(16, 44, 0, 0);
+  superview.customSafeAreaInsets = safeAreaInsets;
+  CGSize fitSize = self.tabBarView.intrinsicContentSize;
+  fitSize = CGSizeMake(fitSize.width + safeAreaInsets.left, fitSize.height + safeAreaInsets.top);
+  self.tabBarView.bounds = CGRectMake(0, 0, fitSize.width, fitSize.height);
   superview.bounds = CGRectMake(0, 0, CGRectGetWidth(self.tabBarView.bounds),
                                 CGRectGetHeight(self.tabBarView.bounds));
   self.tabBarView.center =
@@ -713,8 +731,12 @@ static NSString *const kItemTitleLong3Arabic = @"تحت أي قدما وإقام
   [self.tabBarView setSelectedItem:item2 animated:NO];
 
   // When
-  superview.customSafeAreaInsets = UIEdgeInsetsMake(0, 0, 16, 44);
-  [self.tabBarView sizeToFit];
+  UIEdgeInsets safeAreaInsets = UIEdgeInsetsMake(0, 0, 16, 44);
+  superview.customSafeAreaInsets = safeAreaInsets;
+  CGSize fitSize = self.tabBarView.intrinsicContentSize;
+  fitSize =
+      CGSizeMake(fitSize.width + safeAreaInsets.right, fitSize.height + safeAreaInsets.bottom);
+  self.tabBarView.bounds = CGRectMake(0, 0, fitSize.width, fitSize.height);
   superview.bounds = CGRectMake(0, 0, CGRectGetWidth(self.tabBarView.bounds),
                                 CGRectGetHeight(self.tabBarView.bounds));
   self.tabBarView.center =
@@ -788,6 +810,134 @@ static NSString *const kItemTitleLong3Arabic = @"تحت أي قدما وإقام
 
   // Then
   [self generateSnapshotAndVerifyForView:self.tabBarView];
+}
+
+- (void)testCustomInsetsTopAndRightRepositionsJustifiedViews {
+  // Given
+  UITabBarItem *item1 = [[UITabBarItem alloc] initWithTitle:@"1" image:nil tag:0];
+  UITabBarItem *item2 = [[UITabBarItem alloc] initWithTitle:@"2" image:nil tag:1];
+  self.tabBarView.items = @[ item1, item2 ];
+  [self.tabBarView setSelectedItem:item1 animated:NO];
+
+  // When
+  UIEdgeInsets contentInset = UIEdgeInsetsMake(15, 0, 0, 45);
+  self.tabBarView.contentInset = contentInset;
+  CGSize fitSize = self.tabBarView.intrinsicContentSize;
+  fitSize = CGSizeMake(fitSize.width + contentInset.right, fitSize.height + contentInset.top);
+  self.tabBarView.bounds = CGRectMake(0, 0, fitSize.width, fitSize.height);
+
+  // Then
+  [self generateSnapshotAndVerifyForView:self.tabBarView];
+}
+
+- (void)testCustomInsetsBottomAndLeftRepositionsJustifiedViews {
+  // Given
+  UITabBarItem *item1 = [[UITabBarItem alloc] initWithTitle:@"1" image:nil tag:0];
+  UITabBarItem *item2 = [[UITabBarItem alloc] initWithTitle:@"2" image:nil tag:1];
+  UITabBarItem *item3 = [[UITabBarItem alloc] initWithTitle:@"3" image:nil tag:2];
+  UITabBarItem *item4 = [[UITabBarItem alloc] initWithTitle:@"4" image:nil tag:3];
+  self.tabBarView.items = @[ item1, item2, item3, item4 ];
+  [self.tabBarView setSelectedItem:item1 animated:NO];
+
+  // When
+  UIEdgeInsets contentInset = UIEdgeInsetsMake(0, 45, 15, 0);
+  self.tabBarView.contentInset = contentInset;
+  CGSize fitSize = self.tabBarView.intrinsicContentSize;
+  fitSize = CGSizeMake(fitSize.width + contentInset.left, fitSize.height + contentInset.bottom);
+  self.tabBarView.bounds = CGRectMake(0, 0, fitSize.width, fitSize.height);
+
+  // Then
+  [self generateSnapshotAndVerifyForView:self.tabBarView];
+}
+
+- (void)testCustomInsetsTopAndRightRepositionsScrollableViews {
+  // Given
+  UITabBarItem *item1 = [[UITabBarItem alloc] initWithTitle:@"1" image:nil tag:0];
+  UITabBarItem *item2 = [[UITabBarItem alloc] initWithTitle:@"2" image:nil tag:1];
+  UITabBarItem *item3 = [[UITabBarItem alloc] initWithTitle:@"3" image:nil tag:2];
+  UITabBarItem *item4 = [[UITabBarItem alloc] initWithTitle:@"4" image:nil tag:3];
+  self.tabBarView.items = @[ item1, item2, item3, item4 ];
+  [self.tabBarView setSelectedItem:item4 animated:NO];
+  self.tabBarView.bounds = CGRectMake(0, 0, 120, kExpectedHeightTitlesOrIconsOnly);
+
+  // When
+  self.tabBarView.contentInset = UIEdgeInsetsMake(15, 0, 0, 15);
+
+  // Then
+  [self generateSnapshotAndVerifyForView:self.tabBarView];
+}
+
+- (void)testCustomInsetsBottomAndLeftRepositionsScrollableViews {
+  // Given
+  UITabBarItem *item1 = [[UITabBarItem alloc] initWithTitle:@"1" image:nil tag:0];
+  UITabBarItem *item2 = [[UITabBarItem alloc] initWithTitle:@"2" image:nil tag:1];
+  UITabBarItem *item3 = [[UITabBarItem alloc] initWithTitle:@"3" image:nil tag:2];
+  UITabBarItem *item4 = [[UITabBarItem alloc] initWithTitle:@"4" image:nil tag:3];
+  self.tabBarView.items = @[ item1, item2, item3, item4 ];
+  [self.tabBarView setSelectedItem:item1 animated:NO];
+  self.tabBarView.bounds = CGRectMake(0, 0, 120, kExpectedHeightTitlesOrIconsOnly);
+
+  // When
+  self.tabBarView.contentInset = UIEdgeInsetsMake(0, 15, 15, 0);
+
+  // Then
+  [self generateSnapshotAndVerifyForView:self.tabBarView];
+}
+
+- (void)testSafeAreaAndContentInsetCombineTopAndLeftForJustifiedLayoutStyle {
+  // Given
+  UITabBarItem *item1 = [[UITabBarItem alloc] initWithTitle:@"1" image:nil tag:0];
+  UITabBarItem *item2 = [[UITabBarItem alloc] initWithTitle:@"2" image:nil tag:2];
+  MDCTabBarViewSnapshotTestsSuperview *superview =
+      [[MDCTabBarViewSnapshotTestsSuperview alloc] init];
+  [superview addSubview:self.tabBarView];
+  self.tabBarView.items = @[ item1, item2 ];
+  [self.tabBarView setSelectedItem:item2 animated:NO];
+
+  // When
+  UIEdgeInsets safeAreaInsets = UIEdgeInsetsMake(10, 5, 0, 0);
+  UIEdgeInsets contentInset = UIEdgeInsetsMake(5, 10, 0, 0);
+  superview.customSafeAreaInsets = safeAreaInsets;
+  self.tabBarView.contentInset = contentInset;
+  CGSize fitSize = self.tabBarView.intrinsicContentSize;
+  fitSize = CGSizeMake(fitSize.width + safeAreaInsets.left + contentInset.left,
+                       fitSize.height + safeAreaInsets.top + contentInset.top);
+  self.tabBarView.bounds = CGRectMake(0, 0, fitSize.width, fitSize.height);
+  superview.bounds = CGRectMake(0, 0, CGRectGetWidth(self.tabBarView.bounds),
+                                CGRectGetHeight(self.tabBarView.bounds));
+  self.tabBarView.center =
+      CGPointMake(CGRectGetMidX(superview.bounds), CGRectGetMidY(superview.bounds));
+
+  // Then
+  [self generateSnapshotAndVerifyForView:superview];
+}
+
+- (void)testSafeAreaAndContentInsetCombineBottomAndRightForJustifiedLayoutStyle {
+  // Given
+  UITabBarItem *item1 = [[UITabBarItem alloc] initWithTitle:@"1" image:nil tag:0];
+  UITabBarItem *item2 = [[UITabBarItem alloc] initWithTitle:@"2" image:nil tag:1];
+  MDCTabBarViewSnapshotTestsSuperview *superview =
+      [[MDCTabBarViewSnapshotTestsSuperview alloc] init];
+  [superview addSubview:self.tabBarView];
+  self.tabBarView.items = @[ item1, item2 ];
+  [self.tabBarView setSelectedItem:item2 animated:NO];
+
+  // When
+  UIEdgeInsets safeAreaInsets = UIEdgeInsetsMake(0, 0, 10, 5);
+  UIEdgeInsets contentInset = UIEdgeInsetsMake(0, 0, 5, 10);
+  superview.customSafeAreaInsets = safeAreaInsets;
+  self.tabBarView.contentInset = contentInset;
+  CGSize fitSize = self.tabBarView.intrinsicContentSize;
+  fitSize = CGSizeMake(fitSize.width + safeAreaInsets.right + contentInset.right,
+                       fitSize.height + safeAreaInsets.bottom + contentInset.bottom);
+  self.tabBarView.bounds = CGRectMake(0, 0, fitSize.width, fitSize.height);
+  superview.bounds = CGRectMake(0, 0, CGRectGetWidth(self.tabBarView.bounds),
+                                CGRectGetHeight(self.tabBarView.bounds));
+  self.tabBarView.center =
+      CGPointMake(CGRectGetMidX(superview.bounds), CGRectGetMidY(superview.bounds));
+
+  // Then
+  [self generateSnapshotAndVerifyForView:superview];
 }
 
 #pragma mark - MDCTabBarView Properties
@@ -905,7 +1055,7 @@ static NSString *const kItemTitleLong3Arabic = @"تحت أي قدما وإقام
   // When
   [self.tabBarView setTitleFont:[UIFont systemFontOfSize:8] forState:UIControlStateNormal];
   [self.tabBarView setTitleFont:[UIFont systemFontOfSize:24] forState:UIControlStateSelected];
-  [self.tabBarView sizeToFit];
+  SizeViewToIntrinsicContentSize(self.tabBarView);
 
   // Then
   [self generateSnapshotAndVerifyForView:self.tabBarView];
@@ -921,7 +1071,7 @@ static NSString *const kItemTitleLong3Arabic = @"تحت أي قدما وإقام
 
   // When
   [self.tabBarView setTitleFont:[UIFont systemFontOfSize:8] forState:UIControlStateNormal];
-  [self.tabBarView sizeToFit];
+  SizeViewToIntrinsicContentSize(self.tabBarView);
 
   // Then
   [self generateSnapshotAndVerifyForView:self.tabBarView];
@@ -940,7 +1090,7 @@ static NSString *const kItemTitleLong3Arabic = @"تحت أي قدما وإقام
   // When
   [self.tabBarView setTitleFont:nil forState:UIControlStateNormal];
   [self.tabBarView setTitleFont:nil forState:UIControlStateSelected];
-  [self.tabBarView sizeToFit];
+  SizeViewToIntrinsicContentSize(self.tabBarView);
 
   // Then
   [self generateSnapshotAndVerifyForView:self.tabBarView];
@@ -983,7 +1133,7 @@ static NSString *const kItemTitleLong3Arabic = @"تحت أي قدما وإقام
   UITabBarItem *item3 = [[UITabBarItem alloc] initWithTitle:@"Three" image:self.typicalIcon3 tag:3];
   self.tabBarView.items = @[ item1, item2, item3 ];
   [self.tabBarView setSelectedItem:item2 animated:NO];
-  [self.tabBarView sizeToFit];
+  SizeViewToIntrinsicContentSize(self.tabBarView);
 
   // When
   self.tabBarView.rippleColor = [UIColor.orangeColor colorWithAlphaComponent:(CGFloat)0.25];
@@ -1000,7 +1150,7 @@ static NSString *const kItemTitleLong3Arabic = @"تحت أي قدما وإقام
   UITabBarItem *item3 = [[UITabBarItem alloc] initWithTitle:@"Three" image:self.typicalIcon3 tag:3];
   self.tabBarView.items = @[ item1, item2, item3 ];
   [self.tabBarView setSelectedItem:item2 animated:NO];
-  [self.tabBarView sizeToFit];
+  SizeViewToIntrinsicContentSize(self.tabBarView);
 
   // When
   self.tabBarView.selectionIndicatorStrokeColor = UIColor.purpleColor;
@@ -1016,8 +1166,25 @@ static NSString *const kItemTitleLong3Arabic = @"تحت أي قدما وإقام
   UITabBarItem *item3 = [[UITabBarItem alloc] initWithTitle:@"Three" image:self.typicalIcon3 tag:3];
   self.tabBarView.items = @[ item1, item2, item3 ];
   [self.tabBarView setSelectedItem:item2 animated:NO];
-  [self.tabBarView sizeToFit];
+  SizeViewToIntrinsicContentSize(self.tabBarView);
   self.tabBarView.selectionIndicatorStrokeColor = UIColor.purpleColor;
+
+  // When
+  self.tabBarView.selectionIndicatorStrokeColor = nil;
+
+  // Then
+  [self generateSnapshotAndVerifyForView:self.tabBarView];
+}
+
+- (void)testBottomDividerColor {
+  // Given
+  UITabBarItem *item1 = [[UITabBarItem alloc] initWithTitle:@"One" image:self.typicalIcon1 tag:0];
+  UITabBarItem *item2 = [[UITabBarItem alloc] initWithTitle:@"Two" image:self.typicalIcon2 tag:2];
+  UITabBarItem *item3 = [[UITabBarItem alloc] initWithTitle:@"Three" image:self.typicalIcon3 tag:3];
+  self.tabBarView.items = @[ item1, item2, item3 ];
+  [self.tabBarView setSelectedItem:item2 animated:NO];
+  SizeViewToIntrinsicContentSize(self.tabBarView);
+  self.tabBarView.bottomDividerColor = UIColor.purpleColor;
 
   // When
   self.tabBarView.selectionIndicatorStrokeColor = nil;
@@ -1039,7 +1206,7 @@ static NSString *const kItemTitleLong3Arabic = @"تحت أي قدما وإقام
 
   // When
   self.tabBarView.items = @[ item1, customViewItem, item3 ];
-  [self.tabBarView sizeToFit];
+  SizeViewToIntrinsicContentSize(self.tabBarView);
 
   // Then
   [self generateSnapshotAndVerifyForView:self.tabBarView];